Change input for duration
I would have liked a native input for selecting a duration of time, but
there is no local based time input, so if I wanted to input a duration
of 30 minutes, it would show as 12:30 b/c of the conversion to 12-hour
format. I could have gone with text, but it seems easiest and most
accessible if I just have two different inputs for the hours and
minutes.