Add ability to switch days
I was able to do this earlier by making each day a separate page, but
there would be a FOUC on the map for each new load since it would be
rerendered anew each time, which didn't look nice.
So I moved that into the prop itself, and in the process
refactored a few things for cleanliness.