Move sign up dialog to always visible button
The sign up dialog is connected to a button. This button was hidden
when the user was signed in, which could prove annoying for anonymous
user who want to sign up but could not because they were techincally
logged in.

So I moved the dialog to the ever present call to action button, so the
dialog can always pop up.

Also did some file structure changes while refactoring.